+++ api-base/types/calls.ts
@@ -2,8 +2,19 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { Observable } from 'https://esm.sh/rxjs@7.5.5';
-import type { AnyFunction, Codec, DefinitionCallNamed } from 'https://deno.land/x/polkadot@0.0.3/types/types/index.ts';
+import type { AnyFunction, Codec, DefinitionCallNamed } from 'https://deno.land/x/polkadot/types/types/index.ts';
 import type { ApiTypes, ReturnCodec } from './base.ts';
+
+export type DecoratedCallBase<ApiType extends ApiTypes, F extends AnyFunction = (...args: any[]) => Observable<Codec>> =
+  ApiType extends 'rxjs'
+    ? <T extends Codec | any = ReturnCodec<F>> (...args: Parameters<F>) => Observable<T>
+    : <T extends Codec | any = ReturnCodec<F>> (...args: Parameters<F>) => Promise<T>;
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export type AugmentedCall<ApiType extends ApiTypes, F extends AnyFunction = (...args: any[]) => Observable<Codec>> =
+DecoratedCallBase<ApiType, F> & { meta: DefinitionCallNamed };
+
+// augmented interfaces
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars,@typescript-eslint/no-empty-interface
 export interface AugmentedCalls<ApiType extends ApiTypes> {
@@ -18,12 +29,3 @@ export interface QueryableCalls<ApiType extends ApiTypes> extends AugmentedCalls
 export interface QueryableModuleCalls<ApiType extends ApiTypes> {
   [key: string]: DecoratedCallBase<ApiType>;
 }
-
-export type DecoratedCallBase<ApiType extends ApiTypes, F extends AnyFunction = (...args: any[]) => Observable<Codec>> =
-  ApiType extends 'rxjs'
-    ? <T extends Codec | any = ReturnCodec<F>> (...args: Parameters<F>) => Observable<T>
-    : <T extends Codec | any = ReturnCodec<F>> (...args: Parameters<F>) => Promise<T>;
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export type AugmentedCall<ApiType extends ApiTypes, F extends AnyFunction = (...args: any[]) => Observable<Codec>> =
-DecoratedCallBase<ApiType, F> & { meta: DefinitionCallNamed };