+++ api-base/types/consts.ts
@@ -1,9 +1,16 @@
 // Copyright 2017-2022 @polkadot/api-base authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { PalletConstantMetadataLatest } from 'https://deno.land/x/polkadot@0.0.3/types/interfaces/index.ts';
-import type { Codec } from 'https://deno.land/x/polkadot@0.0.3/types/types/index.ts';
+import type { PalletConstantMetadataLatest } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
+import type { Codec } from 'https://deno.land/x/polkadot/types/types/index.ts';
 import type { ApiTypes } from './base.ts';
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export interface AugmentedConst<ApiType extends ApiTypes> {
+  meta: PalletConstantMetadataLatest;
+}
+
+// augmented interfaces
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars,@typescript-eslint/no-empty-interface
 export interface AugmentedConsts<ApiType extends ApiTypes> {
@@ -13,11 +20,6 @@ export interface AugmentedConsts<ApiType extends ApiTypes> {
 export interface QueryableConsts<ApiType extends ApiTypes> extends AugmentedConsts<ApiType> {
   // when non-augmented, we need to at least have Codec results
   [key: string]: QueryableModuleConsts;
-}
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export interface AugmentedConst<ApiType extends ApiTypes> {
-  meta: PalletConstantMetadataLatest;
 }
 
 export interface QueryableModuleConsts {