+++ api-base/types/errors.ts
@@ -1,8 +1,13 @@
 // Copyright 2017-2022 @polkadot/api-base authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { IsError } from 'https://deno.land/x/polkadot@0.0.3/types/metadata/decorate/types.ts';
+import type { IsError } from 'https://deno.land/x/polkadot/types/metadata/decorate/types.ts';
 import type { ApiTypes } from './base.ts';
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export type AugmentedError<ApiType extends ApiTypes> = IsError;
+
+// augmented interfaces
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars,@typescript-eslint/no-empty-interface
 export interface AugmentedErrors<ApiType extends ApiTypes> {
@@ -13,9 +18,6 @@ export interface DecoratedErrors<ApiType extends ApiTypes> extends AugmentedErro
   // when non-augmented, we need to at least have Codec results
   [key: string]: ModuleErrors<ApiType>;
 }
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export type AugmentedError<ApiType extends ApiTypes> = IsError;
 
 export interface ModuleErrors<ApiType extends ApiTypes> {
   [key: string]: AugmentedError<ApiType>;