+++ api-base/types/events.ts
@@ -1,9 +1,14 @@
 // Copyright 2017-2022 @polkadot/api-base authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { IsEvent } from 'https://deno.land/x/polkadot@0.0.3/types/metadata/decorate/types.ts';
-import type { AnyTuple } from 'https://deno.land/x/polkadot@0.0.3/types/types/index.ts';
+import type { IsEvent } from 'https://deno.land/x/polkadot/types/metadata/decorate/types.ts';
+import type { AnyTuple } from 'https://deno.land/x/polkadot/types/types/index.ts';
 import type { ApiTypes } from './base.ts';
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export type AugmentedEvent<ApiType extends ApiTypes, T extends AnyTuple = AnyTuple, N = unknown> = IsEvent<T, N>;
+
+// augmented interfaces
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars,@typescript-eslint/no-empty-interface
 export interface AugmentedEvents<ApiType extends ApiTypes> {
@@ -14,9 +19,6 @@ export interface DecoratedEvents<ApiType extends ApiTypes> extends AugmentedEven
   // when non-augmented, we need to at least have Codec results
   [key: string]: ModuleEvents<ApiType>;
 }
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export type AugmentedEvent<ApiType extends ApiTypes, T extends AnyTuple = AnyTuple, N = unknown> = IsEvent<T, N>;
 
 export interface ModuleEvents<ApiType extends ApiTypes> {
   [key: string]: AugmentedEvent<ApiType>;