+++ api-base/types/storage.ts
@@ -2,25 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type { Observable } from 'https://esm.sh/rxjs@7.5.5';
-import type { StorageKey, u64 } from 'https://deno.land/x/polkadot@0.0.3/types/mod.ts';
-import type { Hash } from 'https://deno.land/x/polkadot@0.0.3/types/interfaces/index.ts';
-import type { StorageEntry } from 'https://deno.land/x/polkadot@0.0.3/types/primitive/types.ts';
-import type { AnyFunction, AnyTuple, Callback, Codec, IStorageKey } from 'https://deno.land/x/polkadot@0.0.3/types/types/index.ts';
+import type { StorageKey, u64 } from 'https://deno.land/x/polkadot/types/mod.ts';
+import type { Hash } from 'https://deno.land/x/polkadot/types/interfaces/index.ts';
+import type { StorageEntry } from 'https://deno.land/x/polkadot/types/primitive/types.ts';
+import type { AnyFunction, AnyTuple, Callback, Codec, IStorageKey } from 'https://deno.land/x/polkadot/types/types/index.ts';
 import type { ApiTypes, DropLast, MethodResult, PaginationOptions, PromiseOrObs, ReturnCodec, UnsubscribePromise } from './base.ts';
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars,@typescript-eslint/no-empty-interface
-export interface AugmentedQueries<ApiType extends ApiTypes> {
-  // augmented
-}
-
-export interface QueryableStorage<ApiType extends ApiTypes> extends AugmentedQueries<ApiType> {
-  // when non-augmented, we need to at least have Codec results
-  [key: string]: QueryableModuleStorage<ApiType>;
-}
-
-export interface QueryableStorageAt<ApiType extends ApiTypes> extends AugmentedQueries<ApiType> {
-  [key: string]: QueryableModuleStorageAt<ApiType>;
-}
 
 interface StorageEntryObservableMulti<R extends Codec = Codec> {
   <T extends Codec = R>(args: (unknown[] | unknown)[]): Observable<T[]>;
@@ -129,3 +115,19 @@ export type AugmentedQueryAt<ApiType extends ApiTypes, F extends AnyFunction, A 
 
 // backwards compatibility-only
 export type AugmentedQueryDoubleMap<ApiType extends ApiTypes, F extends AnyFunction, A extends AnyTuple = AnyTuple> = AugmentedQuery<ApiType, F, A>;
+
+// augmented interfaces
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars,@typescript-eslint/no-empty-interface
+export interface AugmentedQueries<ApiType extends ApiTypes> {
+  // augmented
+}
+
+export interface QueryableStorage<ApiType extends ApiTypes> extends AugmentedQueries<ApiType> {
+  // when non-augmented, we need to at least have Codec results
+  [key: string]: QueryableModuleStorage<ApiType>;
+}
+
+export interface QueryableStorageAt<ApiType extends ApiTypes> extends AugmentedQueries<ApiType> {
+  [key: string]: QueryableModuleStorageAt<ApiType>;
+}